requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
dvc


# backwards compatibility
pathlib2

# project dependencies
pandas
folium
numpy
scipy
haversine
matplotlib
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
prettytable
pyarrow
fastparquet
openpyxl